# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cahoots API
  description: API for Cahoots AI Development Team Platform
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cahoots.ai
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: string
        customer_id:
          type: string
        price_id:
          type: string
        status:
          type: string
        current_period_end:
          type: string
          format: date-time
        cancel_at_period_end:
          type: boolean
    PaymentMethodResponse:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        card:
          type: object
          properties:
            brand:
              type: string
            last4:
              type: string
            exp_month:
              type: integer
            exp_year:
              type: integer
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
        customer_id:
          type: string
        amount_due:
          type: integer
        amount_paid:
          type: integer
        status:
          type: string
        created:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
    UsageResponse:
      type: object
      properties:
        subscription_id:
          type: string
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        total_usage:
          type: integer
...
paths:
...
  /billing/subscriptions/{subscription_id}:
    parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Billing
      summary: Get subscription details
      operationId: getSubscription
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    patch:
      tags:
        - Billing
      summary: Update subscription
      operationId: updateSubscription
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - price_id
              properties:
                price_id:
                  type: string
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    delete:
      tags:
        - Billing
      summary: Cancel subscription
      operationId: cancelSubscription
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
  /billing/payment-methods:
    get:
      tags:
        - Billing
      summary: List payment methods
      operationId: listPaymentMethods
      security:
        - ApiKeyAuth: []
      parameters:
        - name: customer_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponse'
    post:
      tags:
        - Billing
      summary: Add payment method
      operationId: addPaymentMethod
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_method_id
                - customer_id
              properties:
                payment_method_id:
                  type: string
                customer_id:
                  type: string
      responses:
        '200':
          description: Payment method added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
...
  /billing/invoices:
    get:
      tags:
        - Billing
      summary: List invoices
      operationId: listInvoices
      security:
        - ApiKeyAuth: []
      parameters:
        - name: customer_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'
  /billing/invoices/{invoice_id}:
    parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Billing
      summary: Get invoice details
      operationId: getInvoice
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
  /billing/invoices/{invoice_id}/pay:
    parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - Billing
      summary: Pay invoice
      operationId: payInvoice
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Invoice paid successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
  /billing/usage/{subscription_id}:
    parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Billing
      summary: Get subscription usage
      operationId: getUsage
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Subscription usage details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
...